refactor(api): migrate openapi route module to package layout
