+++ docs/todo.md
@@ -2,7 +2,8 @@
 
 - [x] Verify Team role workflow policy is enforced across prompt defaults and runtime startup gates (leader: `use_existing` + empty workspace; worker: `create_worktree` + isolated run workdir + random branch) with Rust/Web regression tests (see `docs/features/2026-02-20-team-role-workflow-policy.md`).
 - [x] Refactor ACP-related code into package-style module layout (`src/acp/mod.rs` + `event_sink.rs` + `runtime.rs`) and move service bootstrap out of `lib.rs` into `src/app.rs` while preserving `main` thin-entry behavior (`agenthub::run`) and ACP runtime test coverage (see `docs/features/2026-02-20-acp-package-and-bootstrap-module-migration.md`).
-- [x] Refactor OpenAPI spec assembly into dedicated module `src/api/openapi/spec.rs` to keep `src/api/openapi.rs` focused on routing/auth/docs-page handlers, and verify OpenAPI auth/path contract tests remain stable (see `docs/features/2026-02-20-openapi-spec-module-migration.md`).
+- [x] Refactor OpenAPI route module into package layout (`src/api/openapi/mod.rs` + `docs_page.rs` + `tests.rs`) and remove monolithic `src/api/openapi.rs` while keeping `/api/openapi.json` and `/api/openapi/docs` behavior unchanged (see `docs/features/2026-02-20-openapi-package-module-migration.md`).
+- [x] Refactor OpenAPI spec assembly into dedicated module `src/api/openapi/spec.rs` to keep OpenAPI route module focused on routing/auth/docs-page handlers, and verify OpenAPI auth/path contract tests remain stable (see `docs/features/2026-02-20-openapi-spec-module-migration.md`).
 - [ ] Verify startup history retention cleanup (`history.event_retention_days`) deletes `agent_events` older than configured window (default `5` days) and optional `history.vacuum_on_cleanup` compaction behavior remains stable on non-empty cleanup runs (see `docs/features/2026-02-20-sqlite-history-retention-cleanup.md`).
 - [x] Verify `DELETE /api/teams/:id` also cleans Team member agent sessions (`agent_sessions`) by member IDs from Team spec and preserves Team data cascade behavior (`team_runs/steps/events/messages`) (see `docs/features/2026-02-20-team-delete-cleans-member-agent-sessions.md`).
 - [x] Verify Team run workbench readability layout refinement keeps interaction semantics unchanged while reducing stacked density (member health block + run controls/list responsive split) and preserves web checks + mobile Team E2E (see `docs/features/2026-02-20-team-run-panel-readability-layout.md`).
@@ -176,7 +177,7 @@
 - [x] Verify Team delete flow (`DELETE /api/teams/:id` + `/teams` `Delete Team` action) removes team-owned runs/events/steps/mailbox records and keeps next-team selection stable in desktop/mobile browsers (see `docs/features/2026-02-17-team-delete-api-ui.md`).
 - [ ] Verify OpenAPI discovery endpoints (`/api/openapi.json`, `/api/openapi/docs`) auth behavior and docs-page token fetch flow in a real browser (see `docs/features/2026-02-16-openapi-discovery-endpoints.md`).
 - [x] Refactor `web/src/pages/team_page.tsx` into smaller feature-focused components (`TeamSidebar`, `RunPanel`, tab subviews) and a reducer-based state model for maintainability (phase-1: `TeamSidebar` + `TeamRunPanel` + `TeamOverviewPanel` + `TeamEventsPanel` + `TeamMailboxPanel` + `TeamMemberConsolePanel` + `TeamStepsPanel` extracted; reducer phase-1/2/3/4 now cover tab/run filters, run/step controls, mailbox/member conversation state, and create-team wizard/forge state) (see `docs/features/2026-02-16-team-workbench-ui.md`, `docs/features/2026-02-18-team-mailbox-panel-component-extraction.md`, `docs/features/2026-02-18-team-member-console-panel-component-extraction.md`, `docs/features/2026-02-19-team-sidebar-component-extraction.md`, `docs/features/2026-02-19-team-run-panel-component-extraction.md`, `docs/features/2026-02-19-team-overview-events-panel-component-extraction.md`, `docs/features/2026-02-19-team-steps-panel-component-extraction.md`, `docs/features/2026-02-19-team-ui-state-reducer-phase1.md`, `docs/features/2026-02-19-team-ui-state-reducer-phase2-run-step.md`, `docs/features/2026-02-19-team-ui-state-reducer-phase3-mailbox-member.md`, `docs/features/2026-02-19-team-ui-state-reducer-phase4-create-team.md`).
-- [ ] Refactor `src/api/openapi.rs` OpenAPI spec assembly into smaller helpers (`components`, `paths`) or a typed generator approach to reduce single-function complexity (see `docs/features/2026-02-16-openapi-discovery-endpoints.md`).
+- [ ] Refactor `src/api/openapi/spec.rs` OpenAPI spec assembly into smaller helpers (`components`, `paths`) or a typed generator approach to reduce single-function complexity (see `docs/features/2026-02-16-openapi-discovery-endpoints.md`).
 - [x] Add router-level HTTP tests for Team APIs to validate wire payload/status through Axum stack (see `docs/features/2026-02-13-a2a-team-phase1-api-hardening.md`).
 - [x] Add Team spec schema validation and API tests for invalid spec payloads before orchestrator wiring.
 - [x] Define cross-field Team spec constraints (entrypoint/member/step graph) for scheduler phase.
