+++ src/api/openapi/docs_page.rs
@@ -1,42 +1,4 @@
-use axum::{
-    Json, Router,
-    extract::State,
-    http::HeaderMap,
-    response::{Html, IntoResponse},
-    routing::get,
-};
-use serde_json::Value;
-
-use crate::api::authz::require_user;
-use crate::api::error::ApiError;
-use crate::state::AppState;
-
-mod spec;
-
-pub fn router(state: AppState) -> Router {
-    Router::new()
-        .route("/openapi.json", get(get_openapi_json))
-        .route("/openapi/docs", get(get_openapi_docs))
-        .with_state(state)
-}
-
-async fn get_openapi_json(
-    State(state): State<AppState>,
-    headers: HeaderMap,
-) -> Result<Json<Value>, ApiError> {
-    let _user = require_user(&headers, &state).await?;
-    Ok(Json(openapi_spec()))
-}
-
-async fn get_openapi_docs() -> impl IntoResponse {
-    Html(OPENAPI_DOCS_HTML)
-}
-
-fn openapi_spec() -> Value {
-    spec::openapi_spec()
-}
-
-const OPENAPI_DOCS_HTML: &str = r#"<!doctype html>
+pub(super) const OPENAPI_DOCS_HTML: &str = r#"<!doctype html>
 <html lang="en">
   <head>
     <meta charset="utf-8" />
@@ -157,64 +119,3 @@ const OPENAPI_DOCS_HTML: &str = r#"<!doctype html>
   </body>
 </html>
 "#;
-
-#[cfg(test)]
-mod tests {
-    use axum::{
-        body::Body,
-        http::{Method, Request, StatusCode, header},
-    };
-    use serde_json::Value;
-    use tower::util::ServiceExt;
-
-    use crate::api::teams::tests::build_test_state;
-
-    #[tokio::test]
-    async fn openapi_json_requires_authorization() {
-        let state = build_test_state().await;
-        let app = super::router(state);
-
-        let resp = app
-            .oneshot(
-                Request::builder()
-                    .method(Method::GET)
-                    .uri("/openapi.json")
-                    .body(Body::empty())
-                    .expect("build request"),
-            )
-            .await
-            .expect("request openapi without auth");
-        assert_eq!(resp.status(), StatusCode::UNAUTHORIZED);
-    }
-
-    #[tokio::test]
-    async fn openapi_json_contains_team_runs_list_path() {
-        let state = build_test_state().await;
-        let token = crate::api::teams::tests::create_auth_token(&state).await;
-        let app = super::router(state);
-
-        let resp = app
-            .oneshot(
-                Request::builder()
-                    .method(Method::GET)
-                    .uri("/openapi.json")
-                    .header(header::AUTHORIZATION, format!("Bearer {token}"))
-                    .body(Body::empty())
-                    .expect("build authorized request"),
-            )
-            .await
-            .expect("request openapi with auth");
-        assert_eq!(resp.status(), StatusCode::OK);
-
-        let bytes = axum::body::to_bytes(resp.into_body(), usize::MAX)
-            .await
-            .expect("read response body");
-        let value: Value = serde_json::from_slice(&bytes).expect("decode openapi json");
-        assert_eq!(value["openapi"], Value::from("3.0.3"));
-        assert!(value["paths"]["/api/teams/{id}"]["delete"].is_object());
-        assert!(value["paths"]["/api/teams/{id}/runs"].is_object());
-        assert!(value["paths"]["/api/teams/runs/{run_id}/resume"].is_object());
-        assert!(value["paths"]["/api/teams/runs/{run_id}/restart"].is_object());
-        assert!(value["paths"]["/api/teams/runs/{run_id}/snapshot"].is_object());
-    }
-}
