+++ src/api/openapi/mod.rs
@@ -1,0 +1,37 @@
+use axum::{
+    Json, Router,
+    extract::State,
+    http::HeaderMap,
+    response::{Html, IntoResponse},
+    routing::get,
+};
+use serde_json::Value;
+
+use crate::api::authz::require_user;
+use crate::api::error::ApiError;
+use crate::state::AppState;
+
+mod docs_page;
+mod spec;
+
+pub fn router(state: AppState) -> Router {
+    Router::new()
+        .route("/openapi.json", get(get_openapi_json))
+        .route("/openapi/docs", get(get_openapi_docs))
+        .with_state(state)
+}
+
+async fn get_openapi_json(
+    State(state): State<AppState>,
+    headers: HeaderMap,
+) -> Result<Json<Value>, ApiError> {
+    let _user = require_user(&headers, &state).await?;
+    Ok(Json(spec::openapi_spec()))
+}
+
+async fn get_openapi_docs() -> impl IntoResponse {
+    Html(docs_page::OPENAPI_DOCS_HTML)
+}
+
+#[cfg(test)]
+mod tests;
